- Split the recipe for ajtcl-services, because core and services releases can be updated several months apart! - Cleaned the recipes, and generate only one package per recipe - Added AJTCL_CORE_SAMPLES and AJTCL_SERVICE_SAMPLES to control what samples should be added to the packages - Updated the versions for ajtcl and base_tcl to the latest available in master, for _git recipes

+++ recipes-core/alljoyn/ajtcl_15.09.bb
@@ -1,29 +1,35 @@
 SUMMARY = "Alljoyn thin framework and SDK by the Allseen Alliance."
 DESCRIPTION = "Alljoyn is an Open Source framework that makes it easy for devices and apps to discover and securely communicate with each other."
+AUTHOR = "Herve Jourdain <herve.jourdain@beechwoods.com>"
 HOMEPAGE = "https://www.allseenalliance.org/"
-DEPENDS = "openssl libxml2"
 SECTION = "libs"
 LICENSE = "ISC"
+DEPENDS = "openssl libxml2"
 LIC_FILES_CHKSUM = "file://${COREBASE}/meta/files/common-licenses/ISC;md5=f3b90e78ea0cffb20bf5cca7947a896d"
 
 S = "${WORKDIR}/alljoyn"
-SRC_URI = "git://git.allseenalliance.org/gerrit/core/${PN}.git;protocol=https;branch=RB${PV};destsuffix=alljoyn/core/${PN} \
-           git://git.allseenalliance.org/gerrit/services/base_tcl.git;protocol=https;branch=RB${PV};destsuffix=alljoyn/services/base_tcl"
+SRC_URI = "git://git.allseenalliance.org/gerrit/core/${PN}.git;protocol=https;branch=RB${PV};destsuffix=alljoyn/core/ajtcl"
 SRCREV = "${AUTOREV}"
+
+AJTCL_CORE_SAMPLES ?= " \
+                        basic_client \
+                        basic_service \
+                        eventaction_service \
+                        nameChange_client \
+                        net_bus \
+                        SecureClient \
+                        SecureClientECDHE \
+                        SecureService \
+                        SecureServiceECDHE \
+                        signalConsumer_client \
+                        signal_service \
+                      "
 
 AJTCL_BINDIR ?= "/opt/ajtcl/bin"
 
 inherit scons
 
 PACKAGES = " \
-             ${PN}-services-samples-dbg \
-             ${PN}-services-samples \
-             ${PN}-services-dbg \
-             ${PN}-services \
-             ${PN}-services-staticdev \
-             ${PN}-services-dev \
-             ${PN}-samples-dbg \
-             ${PN}-samples \
              ${PN}-dbg \
              ${PN} \
              ${PN}-staticdev \
@@ -40,8 +46,6 @@ do_compile() {
     cd ${S}/core/${PN}
 # GTEST_DIR is required because if gtest framework is present, but GTEST_DIR is not, then it triggers a compilation error!
     scons TARG=linux WS=off GTEST_DIR=${STAGING_DIR_HOST}/${prefix}
-    cd ${S}/services/base_tcl
-    scons TARG=linux WS=off AJ_TCL_ROOT=../../core/ajtcl
     unset CROSS_PREFIX
     unset CROSS_PATH
     unset CROSS_CFLAGS
@@ -51,8 +55,6 @@ do_compile() {
 do_compile_class-native() {
     cd ${S}/core/${PN}
     scons TARG=linux WS=off
-    cd ${S}/services/base_tcl
-    scons WS=off AJ_TCL_ROOT=../../core/ajtcl
 }
 
 do_install() {
@@ -61,64 +63,18 @@ do_install() {
     install ${S}/core/${PN}/dist/lib/* ${D}/${libdir}
     cp -r ${S}/core/${PN}/dist/include/${PN}/* ${D}/${includedir}/${PN}
 # Install ajtcl samples
-    install -d ${D}/${AJTCL_BINDIR}
-    cp -r ${S}/core/${PN}/dist/bin/* ${D}/${AJTCL_BINDIR}
-# Install base_tcl
-    install -d ${D}/${libdir} ${D}/${includedir}/${PN}
-    install ${S}/services/base_tcl/dist/lib/* ${D}/${libdir}
-    cp -r ${S}/services/base_tcl/dist/include/${PN}/* ${D}/${includedir}/${PN}
-# Install base_tcl samples
-    install -d ${D}/${AJTCL_BINDIR}
-    install ${S}/services/base_tcl/dist/bin/* ${D}/${AJTCL_BINDIR}
+    for i in ${AJTCL_CORE_SAMPLES}
+    do
+        if [ -f "${S}/core/${PN}/dist/bin/${i}" ]; then
+            install -d ${D}/${AJTCL_BINDIR}
+            install ${S}/core/${PN}/dist/bin/${i} ${D}/${AJTCL_BINDIR}
+        fi
+    done
 }
 
-FILES_${PN}-services-samples = " \
-                                 ${AJTCL_BINDIR}/NotifConfigSample \
-                               "
-FILES_${PN}-services-samples-dbg = " \
-                                     ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/services/base_tcl/samples/* \
-                                     ${AJTCL_BINDIR}/.debug/NotifConfigSample \
-                                   "
-
-FILES_${PN}-services = " \
-                         ${libdir}/lib${PN}_services.so \
-                       "
-FILES_${PN}-services-dbg = " \
-                             ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/services/base_tcl/src/* \
-                             ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/services/base_tcl/dist/* \
-                             ${libdir}/.debug/lib${PN}_services.so \
-                           "
-FILES_${PN}-services-staticdev = " \
-                                   ${libdir}/lib${PN}_services.a \
-                                 "
-FILES_${PN}-services-dev = " \
-                             ${includedir}/${PN}/services/* \
-                           "
-
-FILES_${PN}-samples-dbg = " \
-                            ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/core/${PN}/samples/* \
-                            ${AJTCL_BINDIR}/.debug/* \
-                          "
-FILES_${PN}-samples = " \
-                        ${AJTCL_BINDIR}/* \
-                      "
 FILES_${PN} = " \
                 ${libdir}/lib${PN}.so \
+                ${AJTCL_BINDIR}/* \
               "
-FILES_${PN}-dbg = " \
-                    ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/core/${PN}/src/* \
-                    ${prefix}/src/debug/${PN}/${PV}-${PR}/alljoyn/core/${PN}/dist/include/${PN}/* \
-                    ${libdir}/.debug/lib${PN}.so \
-                  "
-FILES_${PN}-staticdev = " \
-                          ${libdir}/lib${PN}.a \
-                        "
-FILES_${PN}-dev = " \
-                    ${includedir}/${PN}/* \
-                  "
-
-RDEPENDS_${PN}-samples += "${PN}"
-RDEPENDS_${PN}-services += "${PN}"
-RDEPENDS_${PN}-services-samples += "${PN} ${PN}-services"
 
 BBCLASSEXTEND = "native nativesdk"
